feat(filters): added filters as configuration
