+++ lib/filters/color.js
@@ -1,0 +1,26 @@
+const util = require('util');
+const path = require('path');
+const ColorThief = require('color-thief');
+const onecolor = require('onecolor');
+const fs = require('fs-extra');
+
+const readFile = util.promisify(fs.readFile);
+const colorThief = new ColorThief();
+
+module.exports = async function color(entry, options) {
+	const absolutePath = path.join(entry.basePath, entry.relativePath);
+	let rgb = options.fallbackColor;
+
+	try {
+		const content = await readFile(absolutePath);
+		const [r, g, b] = colorThief.getColor(content);
+
+		rgb = onecolor(`rgb(${r}, ${g}, ${b})`).hex();
+	} catch (e) {
+		// Nope
+	}
+
+	return {
+		color: rgb
+	};
+};
